chore: removes BigQuery writer for risk findings
