+++ server/cmd/gram/deps.go
@@ -15,7 +15,6 @@ import (
 	"strings"
 	"time"
 
-	"cloud.google.com/go/bigquery"
 	"cloud.google.com/go/pubsub/v2"
 	"github.com/ClickHouse/clickhouse-go/v2"
 	"github.com/coreos/go-oidc/v3/oidc"
@@ -56,7 +55,6 @@ import (
 	"github.com/speakeasy-api/gram/server/internal/background"
 	bgtriggers "github.com/speakeasy-api/gram/server/internal/background/triggers"
 	"github.com/speakeasy-api/gram/server/internal/billing"
-	"github.com/speakeasy-api/gram/server/internal/bq"
 	"github.com/speakeasy-api/gram/server/internal/cache"
 	"github.com/speakeasy-api/gram/server/internal/conv"
 	"github.com/speakeasy-api/gram/server/internal/encryption"
@@ -966,28 +964,6 @@ func newPubSubClient(ctx context.Context, c *cli.Context, logger *slog.Logger) (
 	return client, broker, func(context.Context) error { return client.Close() }, nil
 }
 
-func newBigQueryClient(ctx context.Context, c *cli.Context, logger *slog.Logger) (bq.Client, func(ctx context.Context) error, error) {
-	if c.Bool("disable-bigquery-writes") {
-		return bq.NewNoopClient(), noopShutdown, nil
-	}
-
-	// Fall back to auto-detecting the project from ADC / the GKE metadata
-	// server when the flag is unset, mirroring how the Pub/Sub client behaves.
-	// Without this, an empty project ID yields invalid table references and
-	// every insert fails with a googleapi 400.
-	projectID := conv.Default(c.String("gcp-project-id"), bigquery.DetectProjectID)
-	client, err := bigquery.NewClient(ctx, projectID, option.WithLogger(logger.With(attr.SlogComponent("gcp-bigquery-client"))))
-	if err != nil {
-		return nil, noopShutdown, fmt.Errorf("failed to create bigquery client: %w", err)
-	}
-
-	shutdown := func(ctx context.Context) error {
-		return client.Close()
-	}
-
-	return bq.NewClient(client), shutdown, nil
-}
-
 type labelledStop struct {
 	label string
 	pub   interface {
@@ -1055,24 +1031,4 @@ func newPublishers(ctx context.Context, psbroker pubSubBroker) (*background.Publ
 		PromptPolicyAnalysis:    promptPolicyAnalysis,
 		CustomRulesAnalysis:     customRulesAnalysis,
 	}, shutdown, nil
-}
-
-func parseBigQueryTableSpec(spec string) (dataset string, table string, err error) {
-	split := strings.Split(spec, ".")
-	switch len(split) {
-	case 2:
-		return split[0], split[1], nil
-	case 3:
-		return split[1], split[2], nil
-	default:
-		return "", "", fmt.Errorf("invalid BigQuery table spec: %s", spec)
-	}
-}
-
-func bqTableFromSpec(client bq.Client, spec string) (bq.TableHandle, error) {
-	dataset, table, err := parseBigQueryTableSpec(spec)
-	if err != nil {
-		return nil, err
-	}
-	return client.Dataset(dataset).Table(table), nil
 }
